Change cpu governer
Using performance *limits* the cpufreq to the highest non-turbo frequency. The setting should be "ondemand", which will fall back to "powersave" if unavailable. Powersave enables power management on newer intel chips and (as a side effect and contrary to the name) enables turbo modes.

Performance should not be used except for old hardware where power management caused additional latencies.